requirements.txt
boto3==1.9.47
kubernetes==8.0.0
pyOpenSSL==18.0.0
